+++ src/main/kotlin/icu/windea/pls/model/ParadoxRootType.kt
@@ -16,3 +16,4 @@ enum class ParadoxRootType(
 		return description
 	}
 }
+