+++ src/main/kotlin/icu/windea/pls/script/navigation/ParadoxDefinitionChooseByNameContributor.kt
@@ -2,7 +2,7 @@ package icu.windea.pls.script.navigation
 
 import com.intellij.navigation.*
 import com.intellij.psi.search.*
-import com.intellij.psi.stubs.StubIndex
+import com.intellij.psi.stubs.*
 import com.intellij.util.*
 import com.intellij.util.indexing.*
 import icu.windea.pls.script.psi.*
@@ -12,7 +12,7 @@ import icu.windea.pls.script.psi.*
 /**
  * 用于让`Navigate | Class or Navigate | Symbol`可以查找到匹配名字的定义。
  */
-class ParadoxScriptChooseByNameContributor: ChooseByNameContributorEx {
+class ParadoxDefinitionChooseByNameContributor: ChooseByNameContributorEx {
 	override fun processNames(processor: Processor<in String>, scope: GlobalSearchScope, filter: IdFilter?) {
 		StubIndex.getInstance().processAllKeys(ParadoxDefinitionNameIndex.key, processor, scope, filter)
 	}